Make meteorite compass track mysterious cubes
Instead of sky stone blocks.
